defining the backend request to a separate module the backend request to generate the pdf
